Change name of string conversion method to decode
`to_string` implies an implementation of `Display` or `ToString`
for the type, so to disambiguate the method is now called `decode`

I am considering an `Into` implementation, but since the conversion
may fail if the checksum digit fails, `TryInto` would be a better
option.  However, since `TryInto` is only in Nightly, that decision
is pushed down the road.