Add trait encapsulating barcode encode and decode
Generic over formats, allowing support for decoding a variety of
barcode types to a variety of target types, and encoding to formats
from a variety of types.

The ultimate purpose is to allow conversions to rust strings and c
strings in the future, and to support the direct conversion of
arbitrary encoding types directly to a supported Decode type (for
a given format).